requirements.txt
PyYAML
gradio
